Move form into seperate component and migrate styles and tests over
