+++ src/pages/DataItemExplorer/DataItemExplorer.module.scss
@@ -2,24 +2,3 @@
   width: 100%;
   font-family: arial, sans-serif;
 }
-
-.formControls {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-}
-
-.latestValues {
-  margin: 0;
-}
-
-.applicable {
-  display: flex;
-  flex-direction: column;
-}
-
-.buttonsContainer {
-  margin-top: 15px;
-  display: flex;
-  justify-content: space-between;
-}
