+++ src/pages/DataItemExplorer/Form/DataItemExplorerForm.module.scss
@@ -1,0 +1,20 @@
+.formControls {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-end;
+}
+
+.latestValues {
+  margin: 0;
+}
+
+.applicable {
+  display: flex;
+  flex-direction: column;
+}
+
+.buttonsContainer {
+  margin-top: 15px;
+  display: flex;
+  justify-content: space-between;
+}
